// apps/vendor/src/services/models/ReviewViewPagedCollection.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { Link } from './Link';
import type { ReviewView } from './ReviewView';

export type ReviewViewPagedCollection = {
    offset?: number | null;
    limit?: number | null;
    size?: number;
    first?: Link;
    previous?: Link;
    next?: Link;
    last?: Link;
    self?: Link;
    value?: Array<ReviewView> | null;
};